requirements.txt
flask
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
email_validator
wheel
bcrypt
psycopg2-binary
python-dotenv
flask-db
flask-socketio
